+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/OperationContext.java
@@ -48,7 +48,6 @@ public class OperationContext {
   protected String namespace;
   protected boolean defaultNamespace = true;
   protected String name;
-  protected boolean reloadingFromServer;
   protected boolean dryRun;
   protected FieldValidateable.Validation fieldValidation;
   protected String fieldManager;
@@ -74,7 +73,7 @@ public class OperationContext {
   public OperationContext(OperationContext other) {
     this(other.client, other.plural, other.namespace, other.name, other.apiGroupName, other.apiGroupVersion,
         other.item, other.labels, other.labelsNot, other.labelsIn, other.labelsNotIn, other.fields,
-        other.fieldsNot, other.resourceVersion, other.reloadingFromServer, other.gracePeriodSeconds, other.propagationPolicy,
+        other.fieldsNot, other.resourceVersion, other.gracePeriodSeconds, other.propagationPolicy,
         other.dryRun, other.selectorAsString, other.defaultNamespace, other.fieldValidation, other.fieldManager,
         other.forceConflicts);
   }
@@ -82,7 +81,7 @@ public class OperationContext {
   public OperationContext(Client client, String plural, String namespace, String name,
       String apiGroupName, String apiGroupVersion, Object item, Map<String, String> labels,
       Map<String, String[]> labelsNot, Map<String, String[]> labelsIn, Map<String, String[]> labelsNotIn,
-      Map<String, String> fields, Map<String, String[]> fieldsNot, String resourceVersion, boolean reloadingFromServer,
+      Map<String, String> fields, Map<String, String[]> fieldsNot, String resourceVersion,
       long gracePeriodSeconds, DeletionPropagation propagationPolicy,
       boolean dryRun, String selectorAsString, boolean defaultNamespace, FieldValidateable.Validation fieldValidation,
       String fieldManager, Boolean forceConflicts) {
@@ -100,7 +99,6 @@ public class OperationContext {
     setFields(fields);
     setFieldsNot(fieldsNot);
     this.resourceVersion = resourceVersion;
-    this.reloadingFromServer = reloadingFromServer;
     this.gracePeriodSeconds = gracePeriodSeconds;
     this.propagationPolicy = propagationPolicy;
     this.dryRun = dryRun;
@@ -225,10 +223,6 @@ public class OperationContext {
 
   public String getResourceVersion() {
     return resourceVersion;
-  }
-
-  public boolean isReloadingFromServer() {
-    return reloadingFromServer;
   }
 
   public long getGracePeriodSeconds() {
@@ -431,15 +425,6 @@ public class OperationContext {
     return context;
   }
 
-  public OperationContext withReloadingFromServer(boolean reloadingFromServer) {
-    if (this.reloadingFromServer == reloadingFromServer) {
-      return this;
-    }
-    final OperationContext context = new OperationContext(this);
-    context.reloadingFromServer = reloadingFromServer;
-    return context;
-  }
-
   public OperationContext withGracePeriodSeconds(long gracePeriodSeconds) {
     if (this.gracePeriodSeconds == gracePeriodSeconds) {
       return this;
@@ -485,7 +470,7 @@ public class OperationContext {
     // operationcontext
     OperationContext newContext = HasMetadataOperationsImpl.defaultContext(client).withDryRun(getDryRun())
         .withGracePeriodSeconds(getGracePeriodSeconds()).withPropagationPolicy(getPropagationPolicy())
-        .withReloadingFromServer(isReloadingFromServer()).withFieldValidation(this.fieldValidation);
+        .withFieldValidation(this.fieldValidation);
 
     // check before setting to prevent flipping the default flag
     if (!Objects.equals(getNamespace(), newContext.getNamespace())