+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/apps/v1/DeploymentOperationsImpl.java
@@ -89,7 +89,7 @@ public class DeploymentOperationsImpl
     Deployment res = accept(d -> d.getSpec().setReplicas(count));
     if (wait) {
       waitUntilDeploymentIsScaled(count);
-      res = getMandatory();
+      res = getItemOrRequireFromServer();
     }
     return res;
   }